requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.0
backcall==0.2.0
black==23.1.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
debugpy==1.6.7
decorator==5.1.1
Django==4.2
djangorestframework==3.14.0
drf-yasg==1.21.5
executing==1.2.0
firebase-admin==6.1.0
firebase-auth==3.0.28
flex==6.14.1
frozenlist==1.3.3
gcloud==0.17.0
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
inflection==0.5.1
ipykernel==6.22.0
ipython==8.12.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jws==0.1.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
oauth2client==3.0.0
openai==0.27.4
packaging==23.0
parso==0.8.3
pathspec==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.1.1
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.22.1
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.10.1
Pygments==2.15.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jwt==2.0.1
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.2
requests==2.28.2
requests-toolbelt==0.10.1
rfc3987==1.3.8
rsa==4.9
ruamel.yaml==0.17.21
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
strict-rfc3339==0.7
tomli==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
uritemplate==4.1.1
urllib3==1.26.15
validate-email==1.3
wcwidth==0.2.6
yarl==1.8.2
